chore(package.json): update @smartb/g2 and @smartb/g2-i2-v2 versions to 1.0.0-alpha.169 fix(keycloakRequest.ts): refactor postForm function to use DOM manipulation instead of @smartb/g2 request function fix(Login.tsx): add loading state to login button and set isAuthenticating state to true before form submission and false after The package.json file was updated to use the latest versions of @smartb/g2 and @smartb/g2-i2-v2 packages, specifically version 1.0.0-alpha.169. This update may include bug fixes or new features.
In the keycloakRequest.ts file, the postForm function was refactored to use DOM manipulation instead of the @smartb/g2 request function. This change removes the dependency on @smartb/g2 and simplifies the code.

In the Login.tsx file, a loading state was added to the login button. The isAuthenticating state is set to true before form submission and false after. This provides visual feedback to the user that the login process is in progress.