fix(dev.yml): remove docker-buildx-platform from dev workflow to simplify the configuration refactor(Makefile): remove docker push command to avoid pushing images automatically style(Dockerfile): uncomment NODE_OPTIONS environment variable for clarity and consistency refactor(auth.ts, roles.ts): update import paths for domain packages to reflect new package names refactor(GetOrganizationRefsQuery.ts, useOrganizationFormState.ts): update import paths for configuration to reflect new package names refactor(index.ts): update import path for organization domain package to reflect new package name refactor(useUserFormState.ts, useUserColumns.ts): update import paths for configuration to reflect new package names refactor(index.ts): update import path for user domain package to reflect new package name refactor(add.ts, remove.ts, model.ts, get.ts): update import paths for apikey domain package to reflect new package name
The changes were made to update import paths and configurations to reflect the new package names and improve consistency in the codebase.

fix(api): update import path for io module in Apikeys page.ts and OrganizationForm.tsx to reflect the correct domain
fix(components): update property name in useAPIKeyColumn and useFileListColumn functions to match the expected value
fix(config): update import path for i2Config to imConfig in useAuthenticatedRequest.tsx and useUserFunctionnalities.tsx to use the correct configuration

The changes were made to ensure that the correct import paths are used for the io module in the Apikeys and Organization components. Additionally, the property names in the useAPIKeyColumn and useFileListColumn functions were updated to match the expected values. Lastly, the import path for the configuration module was updated from i2Config to imConfig in the useAuthenticatedRequest and useUserFunctionnalities functions for consistency and correctness.