chore(env_dev): update version numbers to stable releases fix(docker-compose-fs.yml): change MINIO_SERVER_URL to use service name chore(package.json): update @komune-io dependencies to version 0.18.0
Updating version numbers from snapshot to stable releases ensures that the application uses stable and tested versions. Changing the MINIO_SERVER_URL to use the service name 'minio' instead of a variable improves reliability in the Docker environment. Updating the @komune-io dependencies to the latest version ensures compatibility and access to the latest features and fixes.