chore(env): update VERSION_CONNECT to 0.21.0-SNAPSHOT and           CONNECT_ADMIN_KEYCLOAK_VERSION to 0.2.0-SNAPSHOT chore(keycloak): modify im-config.json to change "space" to "identifier"                  and add login settings chore(docker): update Dockerfile to copy the new theme structure and                 add additional build steps chore(ci): remove obsolete CI workflow for keycloak
The changes update versioning for the connect admin keycloak and improve
the configuration by changing the "space" field to "identifier" in the
im-config.json file. Additional login settings are added to enhance
functionality. The Dockerfile is modified to accommodate the new theme
structure and build process. The obsolete CI workflow is removed to
streamline the project.